requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
geopy==2.4.1
bcrypt==4.1.2
alembic
requests